individual/requirements.txt
requests
beautifulsoup4
pandas
pymavlink
matplotlib
seaborn
scikit-learn
apache-airflow
